Work only with 1/3 of training data and use only 50% for training set.
In order to speeep up model fitting and make it possible to decide how many epochs we need I included the following changes:
1. Only work with the first  156060/3= 52020 lines of the training data set.
2. In each epoch, the model will randomly split the data into a training and a validation set.
Advantages: The model is much faster and we can check in the log statements when the
accuracy in the validation set stabilizes. Then we can stop the simulation
because the model overfits!